Move winnowing to assembly
